# fix: infoview performance