--- a/lean4-infoview/src/infoview/goalLocation.tsx
+++ b/lean4-infoview/src/infoview/goalLocation.tsx
@@ -5,7 +5,6 @@ Authors: E.W.Ayers, Wojciech Nawrocki
 */
 import { FVarId, MVarId, SubexprPos } from '@leanprover/infoview-api'
 import * as React from 'react'
-import { HoverState } from './tooltips'
 
 /**
  * A location within a goal. It is either:
@@ -76,102 +75,34 @@ export interface Locations {
 
 export const LocationsContext = React.createContext<Locations | undefined>(undefined)
 
-export type HoverSettings = { highlightOnHover: true } | { highlightOnHover: false }
-export type ModHoverSettings = { highlightOnModHover: true } | { highlightOnModHover: false }
-export type SelectionSettings = { highlightOnSelection: true; loc: GoalsLocation } | { highlightOnSelection: false }
-export interface LocationHighlightSettings {
-    ref: React.RefObject<HTMLSpanElement>
-    /**
-     * Whether the span should be highlighted on hover.
-     */
-    hoverSettings: HoverSettings
-    /**
-     * Whether the span should be highlighted on hover while holding `Ctrl` / `Meta`.
-     */
-    modHoverSettings: ModHoverSettings
-    /**
-     * Whether the span should be highlighted when selected, and the current {@link GoalsLocation} that is selected.
-     */
-    selectionSettings: SelectionSettings
-}
-export interface HighlightedLocation {
-    hoverState: HoverState
-    setHoverState: React.Dispatch<React.SetStateAction<HoverState>>
+export type SelectableLocationSettings =
+    | {
+          isSelectable: true
+          /** The {@link GoalsLocation} that can be selected by interacting with the span. */
+          loc: GoalsLocation
+      }
+    | { isSelectable: false }
+
+export interface SelectableLocation {
     className: string
-    onPointerEvent: (b: boolean, e: React.PointerEvent<HTMLSpanElement>) => void
-    onPointerMove: (e: React.PointerEvent<HTMLSpanElement>) => void
-    onKeyDown: (e: React.KeyboardEvent<HTMLSpanElement>) => void
-    onKeyUp: (e: React.KeyboardEvent<HTMLSpanElement>) => void
+    /** Returns whether propagation of the click event within the same handler should be stopped. */
     onClick: (e: React.MouseEvent<HTMLSpanElement, MouseEvent>) => boolean
     onPointerDown: (e: React.PointerEvent<HTMLSpanElement>) => void
 }
 
-/**
- * Logic for a `<span>` with a corresponding {@link GoalsLocation} that can be (un)selected using shift-click.
- */
-export function useHighlightedLocation(settings: LocationHighlightSettings): HighlightedLocation {
-    const { ref, hoverSettings, modHoverSettings, selectionSettings } = settings
-
-    const [hoverState, setHoverState] = React.useState<HoverState>('off')
-
+/** Logic for a span that can be selected using Shift-click and is highlighted when selected. */
+export function useSelectableLocation(settings: SelectableLocationSettings): SelectableLocation {
     const locs = React.useContext(LocationsContext)
     let className: string = ''
-    if (hoverSettings.highlightOnHover && hoverState !== 'off') {
-        className += 'highlight '
-    } else if (selectionSettings.highlightOnSelection && locs && locs.isSelected(selectionSettings.loc)) {
+    if (settings.isSelectable && locs && locs.isSelected(settings.loc)) {
         className += 'highlight-selected '
-    }
-    if (modHoverSettings.highlightOnModHover && hoverState === 'ctrlOver') {
-        className += 'underline '
-    }
-
-    const onPointerEvent = (b: boolean, e: React.PointerEvent<HTMLSpanElement>) => {
-        // It's more composable to let pointer events bubble up rather than to call `stopPropagation`,
-        // but we only want to handle hovers in the innermost component. So we record that the
-        // event was handled with a property.
-        // The `contains` check ensures that the node hovered over is a child in the DOM
-        // tree and not just a logical React child (see useLogicalDom and
-        // https://reactjs.org/docs/portals.html#event-bubbling-through-portals).
-        if (ref.current && e.target instanceof Node && ref.current.contains(e.target)) {
-            if ('_DetectHoverSpanSeen' in e) {
-                return
-            }
-            ;(e as any)._DetectHoverSpanSeen = {}
-            if (!b) {
-                setHoverState('off')
-            } else if (e.ctrlKey || e.metaKey) {
-                setHoverState('ctrlOver')
-            } else {
-                setHoverState('over')
-            }
-        }
-    }
-
-    const onPointerMove = (e: React.PointerEvent<HTMLSpanElement>) => {
-        if (e.ctrlKey || e.metaKey) {
-            setHoverState(st => (st === 'over' ? 'ctrlOver' : st))
-        } else {
-            setHoverState(st => (st === 'ctrlOver' ? 'over' : st))
-        }
-    }
-
-    const onKeyDown = (e: React.KeyboardEvent<HTMLSpanElement>) => {
-        if (e.key === 'Control' || e.key === 'Meta') {
-            setHoverState(st => (st === 'over' ? 'ctrlOver' : st))
-        }
-    }
-
-    const onKeyUp = (e: React.KeyboardEvent<HTMLSpanElement>) => {
-        if (e.key === 'Control' || e.key === 'Meta') {
-            setHoverState(st => (st === 'ctrlOver' ? 'over' : st))
-        }
     }
 
     const onClick = (e: React.MouseEvent<HTMLSpanElement, MouseEvent>) => {
         // On shift-click, if we are in a context where selecting subexpressions makes sense,
         // (un)select the current subexpression.
-        if (selectionSettings.highlightOnSelection && locs && e.shiftKey) {
-            locs.setSelected(selectionSettings.loc, on => !on)
+        if (settings.isSelectable && locs && e.shiftKey) {
+            locs.setSelected(settings.loc, on => !on)
             e.stopPropagation()
             return true
         }
@@ -181,19 +112,13 @@ export function useHighlightedLocation(settings: LocationHighlightSettings): Hig
     const onPointerDown = (e: React.PointerEvent<HTMLSpanElement>) => {
         // We have special handling for shift+click events, so prevent default browser
         // events from interfering when shift is held.
-        if (selectionSettings.highlightOnSelection && locs && e.shiftKey) {
+        if (settings.isSelectable && locs && e.shiftKey) {
             e.preventDefault()
         }
     }
 
     return {
-        hoverState,
-        setHoverState,
         className,
-        onPointerEvent,
-        onPointerMove,
-        onKeyDown,
-        onKeyUp,
         onClick,
         onPointerDown,
     }