… lean4-infoview/package.json
… lean4-infoview/src/infoview/info.tsx
… lean4-infoview/src/infoview/interactiveCode.tsx
… lean4-infoview/src/infoview/messages.tsx
… lean4-infoview/src/infoview/tooltips.tsx
… lean4-infoview/src/infoview/util.ts
… vscode-lean4/src/infoview.ts
… vscode-lean4/test/suite/info/info.test.ts