@mhuisi 2024-07-19
@Vtec234 
> Btw, the selection highlight on parenthesized terms seems to appear twice (not a regression, it was like that before). Do you have a sense of whether the fix is easy, since you're already touching that code?

This is because both a parenthesized subexpression `(foo)` and its subexpression `foo` have the same subexpression position. I made a draft PR to change this at lean4#4786, since it would also help cut down the size of the DOM further. Do you see any obvious issues with this?