@mhuisi 2024-07-19
I'm afraid that this version is slowing down the InfoView by 150ms again. I'll look into why.