@mhuisi 2024-07-19
I suspect the reason was that memoizing in the recursion itself is inhibiting some React optimization. For an experiment, I also tried replacing the recursive component with a recursive function that builds the component, but this did not yield a speed-up.

For some reason I cannot explain, this version is sometimes still 50ms slower than the version at [433fa89](https://github.com/leanprover/vscode-lean4/pull/500/commits/433fa8928f93fd992837393b6b0bcdcde9207bd7) with `genericMemo`. I don't think it's the extra component wrapper because I tried to eliminate that and it was still slower, whereas the version with `genericMemo` consistently seems to be a bit faster.