@Vtec234 2024-07-20
> this version is sometimes still 50ms slower than the version at 433fa89 with `genericMemo`

Hm, there have been other changes (ctrl-related) since that commit. Does a direct comparison with those included result in any difference?