@mhuisi 2024-07-20
No, I double-checked that this is not caused by the new effects. In any case, I think it's fine to go ahead even if it is (in my profiler measurements) sometimes a tiny bit slower.